Small refactor
The function's parameters weren't being utilized. Since they give the script greater flexibility I removed the specific numbers from the function's if statement and replaced them with the parameter values.

The function can be called directly and printed in one statement without the need to assign it to a variable. With the parameters being utilized in the function the numbers being used can be changed with each call of the function.